refactor(stepper): move onboarding-progress to shared step components dir
The indicator is the onboarding progress shared across the onboarding
domain/plans steps and checkout, not a domain-flow component. Moves it
from flows/domain/components to the shared
internals/steps-repository/components library and updates the three
import sites.
