+++ client/landing/stepper/declarative-flow/internals/steps-repository/components/onboarding-progress/index.tsx
