+++ client/landing/stepper/declarative-flow/internals/steps-repository/components/onboarding-progress/test/use-show-onboarding-progress.test.tsx
